smart_contracts/requirements.txt
pyteal>=0.25.0
py-algorand-sdk>=2.4.0
python-dotenv>=1.0.0
